+++ web/apps/cast/src/pages/index.tsx
@@ -3,7 +3,7 @@ import EnteSpinner from "@ente/shared/components/EnteSpinner";
 import LargeType from "components/LargeType";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
-import { advertiseCode, getCastData, register } from "services/cast";
+import { advertiseCode, getCastData, register } from "services/pair";
 import { storeCastData } from "services/cast/castService";
 import { useCastReceiver } from "../utils/useCastReceiver";
